requirements.txt
sentence-transformers==3.0.1
faiss-cpu==1.8.0
langchain-google-genai==2.0.5
langchain-core==0.3.15
numpy==1.26.4
torch==2.3.1
transformers==4.44.2
huggingface-hub==0.24.7
tokenizers==0.19.1
